fix/login: Ban rerendings of the movie poster
